# Reduce network pressure on multi row insert