--- a/go/test/endtoend/vtgate/queries/benchmark/sharded_schema.sql
+++ b/go/test/endtoend/vtgate/queries/benchmark/sharded_schema.sql
@@ -1,0 +1,16 @@
+create table tbl_no_lkp_vdx
+(
+    id  bigint,
+    c1 varchar(50),
+    c2 varchar(50),
+    c3 varchar(50),
+    c4 varchar(50),
+    c5 varchar(50),
+    c6 varchar(50),
+    c7 varchar(50),
+    c8 varchar(50),
+    c9 varchar(50),
+    c10 varchar(50),
+    c11 varchar(50),
+    c12 varchar(50)
+) Engine = InnoDB;