--- a/go/vt/vtgate/engine/insert.go
+++ b/go/vt/vtgate/engine/insert.go
@@ -25,6 +25,7 @@ import (
 	"sync"
 	"time"
 
+	"vitess.io/vitess/go/slice"
 	"vitess.io/vitess/go/sqltypes"
 	"vitess.io/vitess/go/vt/key"
 	querypb "vitess.io/vitess/go/vt/proto/query"
@@ -75,7 +76,7 @@ type (
 
 		// Prefix, Mid and Suffix are for sharded insert plans.
 		Prefix string
-		Mid    []string
+		Mid    sqlparser.Values
 		Suffix string
 
 		// Option to override the standard behavior and allow a multi-shard insert
@@ -132,7 +133,7 @@ func NewInsert(
 	vindexValues [][][]evalengine.Expr,
 	table *vindexes.Table,
 	prefix string,
-	mid []string,
+	mid sqlparser.Values,
 	suffix string,
 ) *Insert {
 	ins := &Insert{
@@ -758,17 +759,23 @@ func (ins *Insert) getInsertShardedRoute(
 
 	queries := make([]*querypb.BoundQuery, len(rss))
 	for i := range rss {
+		shardBindVars := map[string]*querypb.BindVariable{}
 		var mids []string
 		for _, indexValue := range indexesPerRss[i] {
 			index, _ := strconv.ParseInt(string(indexValue.Value), 0, 64)
 			if keyspaceIDs[index] != nil {
-				mids = append(mids, ins.Mid[index])
+				mids = append(mids, sqlparser.String(ins.Mid[index]))
+				for _, expr := range ins.Mid[index] {
+					if arg, ok := expr.(*sqlparser.Argument); ok {
+						shardBindVars[arg.Name] = bindVars[arg.Name]
+					}
+				}
 			}
 		}
 		rewritten := ins.Prefix + strings.Join(mids, ",") + ins.Suffix
 		queries[i] = &querypb.BoundQuery{
 			Sql:           rewritten,
-			BindVariables: bindVars,
+			BindVariables: shardBindVars,
 		}
 	}
 
@@ -985,7 +992,10 @@ func (ins *Insert) description() PrimitiveDescription {
 		other["VindexOffsetFromSelect"] = valuesOffsets
 	}
 	if len(ins.Mid) > 0 {
-		shardQuery := fmt.Sprintf("%s%s%s", ins.Prefix, strings.Join(ins.Mid, ", "), ins.Suffix)
+		mids := slice.Map(ins.Mid, func(from sqlparser.ValTuple) string {
+			return sqlparser.String(from)
+		})
+		shardQuery := fmt.Sprintf("%s%s%s", ins.Prefix, strings.Join(mids, ", "), ins.Suffix)
 		if shardQuery != ins.Query {
 			other["ShardedQuery"] = shardQuery
 		}