… go/vt/vtgate/engine/insert_test.go
… go/vt/vtgate/executor_vschema_ddl_test.go